initial LaTeX support (resolves #1)
This adds support for LaTeX rendering using KaTeX, a fast and popular
library. Since this is a custom extension to Markdown, all of this is
non-standard and doesn't work 100%, e.g. if you include math in some
<pre> code, it will get converted to HTML there (and shown to the user).

But otherwise it's pretty good, just use $ or $$ for inline or
block-based math, respectively.